feat: 공지사항 접근권한 구현

+++ app-server/src/main/java/bitcamp/myapp/controller/AnnouncementController.java
@@ -4,6 +4,8 @@ import bitcamp.myapp.service.NcpObjectStorageService;
 import bitcamp.myapp.service.AnnouncementService;
 import bitcamp.myapp.vo.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -22,30 +24,38 @@ public class AnnouncementController {
   @Autowired
   NcpObjectStorageService ncpObjectStorageService;
 
-
   @GetMapping("form")
-  public void form(@RequestParam("currentPage") int currentPage, Model model) {
+  public void form(
+          @RequestParam("currentPage") int currentPage,
+          Model model,
+          HttpSession session
+  ) throws Exception {
     model.addAttribute("currentPage", currentPage);
 
+    User loginUser = (User) session.getAttribute("loginUser");
+    if (loginUser == null || loginUser.getAuthority() != Authority.ADMIN) {
+      throw new Exception("로그인이 되어있지 않거나 권한이 없습니다.");
+    }
   }
 
   @ResponseBody
   @PostMapping("add")
-  public boolean addAnnouncement(
-//          @RequestParam("currentPage") int currentPage,
+  public ResponseEntity<Object> addAnnouncement(
+          @RequestParam("currentPage") int currentPage,
           Announcement requestData,
           MultipartFile[] files,
           HttpSession session,
           Model model
   ) throws Exception {
-//    System.out.println("announcementController.add 호출");
+    model.addAttribute("currentPage", currentPage);
+
     User loginUser = (User) session.getAttribute("loginUser");
-    if (loginUser == null) {
-      throw new RuntimeException("로그인이 되어있지 않습니다.");
+    if(loginUser == null) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("NoLogin");
     }
-//    else if (!loginUser.getAuthority().equals("ADMIN")) {
-//      return "redirect:/announcement/list?currentPage=" + currentPage;
-//    }
+    if (loginUser.getAuthority() != Authority.ADMIN) {
+      return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("NoAdmin");
+    }
 
     ArrayList<AnnouncementAttachedFile> announcementAttachedFiles = new ArrayList<>();
     for (MultipartFile part : files) {
@@ -61,67 +71,42 @@ public class AnnouncementController {
     requestData.setAnnouncementAttachedFiles(announcementAttachedFiles);
 
     if (announcementService.add(requestData) == 0) {
-      return false;
+      return ResponseEntity.ok(false);
     }
-
-    return true;
+    return ResponseEntity.ok(true);
   }
-
-
-//  @PostMapping("add")
-//  public String add(@RequestParam("currentPage") int currentPage, Announcement announcement,
-//                    MultipartFile[] files, HttpSession session, Model model) throws Exception {
-//
-//    System.out.println("AnnouncementController.add 호출");
-//    System.out.println("AnnouncementController.add / currentPage=  " + currentPage);
-//
-//    User loginUser = (User) session.getAttribute("loginUser");
-//    if (loginUser == null) {
-//      return "redirect:/auth/form";
-//    }
-//
-//    ArrayList<AnnouncementAttachedFile> announcementAttachedFiles = new ArrayList<>();
-//    for (MultipartFile part : files) {
-//      if (part.getSize() > 0) {
-//        String uploadFileUrl = ncpObjectStorageService.uploadFile(
-//                "bitgallery", "announcement/", part);
-//        AnnouncementAttachedFile announcementAttachedFile = new AnnouncementAttachedFile();
-//        announcementAttachedFile.setFilePath(uploadFileUrl);
-//        announcementAttachedFiles.add(announcementAttachedFile);
-//      }
-//    }
-//    announcement.setWriter(loginUser);
-//    announcement.setAnnouncementAttachedFiles(announcementAttachedFiles);
-//
-//    if (announcementService.add(announcement) == 0) {
-//      return "redirect:/announcement/form?currentPage=" + currentPage;
-//    }
-//
-//    return "redirect:/announcement/list?currentPage=" + 1;
-//  }
 
   @GetMapping("delete")
   public String delete(@RequestParam("currentPage") int currentPage, int no, HttpSession session) throws Exception {
+    Announcement announcement = announcementService.get(no);
     User loginUser = (User) session.getAttribute("loginUser");
-    if (loginUser == null) {
-      return "redirect:/auth/form";
-    }
 
-    Announcement a = announcementService.get(no);
-
-    if (a == null || a.getWriter().getNo() != loginUser.getNo()) {
-      throw new Exception("해당 번호의 게시글이 없거나 삭제 권한이 없습니다.");
+    if (loginUser == null || loginUser.getAuthority() != Authority.ADMIN) {
+      throw new Exception("로그인이 되어있지 않거나 권한이 없습니다.");
+    } else if (announcement == null) {
+      throw new Exception("해당 번호의 게시물이 없습니다.");
     } else {
-      announcementService.delete(a.getNo());
+      announcementService.delete(announcement.getNo());
       return "redirect:/announcement/list?currentPage=" + currentPage;
     }
   }
 
   @GetMapping("detail")
   public String detail(@RequestParam("isEdit") boolean isEdit, @RequestParam("currentPage") int currentPage,
-                       @RequestParam("no") int no, Model model) throws Exception {
+                       @RequestParam("no") int no, Model model, HttpSession session) throws Exception {
+
+    User loginUser = (User) session.getAttribute("loginUser");
+    if (loginUser == null) {
+      model.addAttribute("authority", "User");
+      throw new Exception("로그인이 되어있지 않습니다.");
+    } else if (loginUser.getAuthority() != Authority.ADMIN) {
+      throw new Exception("권한이 없습니다.");
+    } else {
+      model.addAttribute("authority", loginUser.getAuthority());
+    }
+
     model.addAttribute("isEdit", isEdit);
-    System.out.println("isEdit =  " + isEdit);
+//    System.out.println("isEdit =  " + isEdit);
     model.addAttribute("currentPage", currentPage);
 
     Announcement announcement = announcementService.get(no);
@@ -133,8 +118,16 @@ public class AnnouncementController {
   }
 
   @GetMapping("list")
-  public void list(@RequestParam("currentPage") int currentPage, Model model) throws Exception {
+  public void list(@RequestParam("currentPage") int currentPage, Model model, HttpSession session) throws Exception {
     model.addAttribute("currentPage", currentPage);
+
+    User loginUser = (User) session.getAttribute("loginUser");
+    if (loginUser == null) {
+      model.addAttribute("authority", "User");
+    } else {
+      model.addAttribute("authority", loginUser.getAuthority());
+    }
+
     model.addAttribute("test", 1);
     announcementService.list(model);
   }
@@ -142,14 +135,12 @@ public class AnnouncementController {
   @PostMapping("update")
   public String update(@RequestParam("currentPage") int currentPage, Announcement announcement,
                        MultipartFile[] files, HttpSession session) throws Exception {
-    System.out.println(currentPage + "업데이트 호출됨");
-    User loginUser = (User) session.getAttribute("loginUser");
-    if (loginUser == null) {
-      return "redirect:/auth/form";
-    }
+//    System.out.println(currentPage + "업데이트 호출됨");
     Announcement a = announcementService.get(announcement.getNo());
-    if (a == null || a.getWriter().getNo() != loginUser.getNo()) {
-      throw new Exception("게시글이 존재하지 않거나 변경 권한이 없습니다.");
+    User loginUser = (User) session.getAttribute("loginUser");
+
+    if (loginUser == null || loginUser.getAuthority() != Authority.ADMIN) {
+      throw new Exception("로그인이 되어있지 않거나 권한이 없습니다.");
     }
 
     ArrayList<AnnouncementAttachedFile> announcementAttachedFiles = new ArrayList<>();
@@ -171,19 +162,19 @@ public class AnnouncementController {
   @GetMapping("fileDelete")
   public String fileDelete(@RequestParam("isEdit") boolean isEdit, @RequestParam("currentPage") int currentPage,
                            @RequestParam("no") int no, HttpSession session, Model model) throws Exception {
-    model.addAttribute("isEdit", isEdit);
-    System.out.println("AnnouncementController.fileDelete.isEdit =  " + isEdit);
     User loginUser = (User) session.getAttribute("loginUser");
-    if (loginUser == null) {
-      return "redirect:/auth/form";
+    if (loginUser == null || loginUser.getAuthority() != Authority.ADMIN) {
+      throw new Exception("로그인이 되어있지 않거나 권한이 없습니다.");
     }
+
+    model.addAttribute("isEdit", isEdit);
+//    System.out.println("AnnouncementController.fileDelete.isEdit =  " + isEdit);
+
     Announcement announcement = null;
     AnnouncementAttachedFile announcementAttachedFile = announcementService.getAnnouncementAttachedFile(no);
     announcement = announcementService.get(announcementAttachedFile.getAnnouncementNo());
-    System.out.println(no + "  no 호출!");
-    if (announcement.getWriter().getNo() != loginUser.getNo()) {
-      throw new Exception("게시글 변경 권한이 없습니다!");
-    }
+//    System.out.println(no + "  no 호출!");
+
     if (announcementService.deleteAttachedFile(no) == 0) {
       throw new Exception("해당 번호의 첨부파일이 없습니다.");
     } else {
